requirements.txt
numpy
torch
pandas
plotly
dash
dash-html-components
dash-core-components
dash-renderer
PyYAML
